+++ .gitignore
@@ -30,5 +30,6 @@ yarn-error.log*
 .env.test.local
 .env.production.local
 
-# vercel
+# deploy
 .vercel
+.serverless_nextjs
