+++ package.json
@@ -7,6 +7,7 @@
     "dev": "cross-env NODE_ENV=development next -p 3060",
     "build": "cross-env ANALYZE=true NODE_ENV=production next build",
     "start": "cross-env NODE_ENV=production next start -p 3060",
+    "deploy": "serverless",
     "analyze": "cross-env ANALYZE=true next build"
   },
   "keywords": [],
@@ -27,6 +28,7 @@
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "sass": "^1.32.6",
+    "serverless": "^2.35.0",
     "tailwindcss": "^2.0.3"
   },
   "devDependencies": {
