+++ serverless.yml
@@ -1,9 +1,8 @@
 slsNextApp:
   component: '@sls-next/serverless-component@1.18.0'
   inputs:
-	  bucketName: 'nodelab.mixin'
-	  cloudfront:
-	    distributionId: 'EOL6OV62VGD1C'
-	  name:
-	    defaultLambda: 'mixin-default-lambda'
-	    apiLambda: 'mixin-api-lambda'
+    bucketName: 'mixin-front2'
+    domain: ["www", "mix-in.net"]
+    name:
+      defaultLambda: 'mixin-front-lambda'
+      apiLambda: 'mixin-front-api-lambda'
